Expose native web terminal close notifications
Report transport close details before rejecting pre-frame mounts without interpreting workload completion or reconnecting automatically.


Copilot-Session: d191512e-7810-45c0-9318-0bbddc20b608
