+++ src/web-terminal/README.md
@@ -51,6 +51,51 @@ terminal.focus();
 or disposes a mounted view. Disposal removes only the appended element and its
 connection, not the container or server-side shared terminal.
 
+### Connection closure and workload completion
+
+Use `onClose(details)` to observe the browser's actual WebSocket close event.
+`TerminalCloseDetails` contains readonly `code`, `reason`, and `wasClean` fields.
+The callback runs once with the view already disconnected, **even if the socket
+closes before the first HWT frame or authoritative HMP peer state**. A pending
+mount rejects after the callback, so capture any host state before calling mount.
+The details object is frozen. Reasons are untrusted text; do not render them as HTML.
+
+```ts
+import { WebTerminal, type TerminalCloseDetails } from "@hex1b/web-terminal";
+
+const container = document.getElementById("terminal");
+if (!container) throw new Error("Missing terminal container");
+let closed: TerminalCloseDetails | undefined;
+try {
+  const terminal = await WebTerminal.mount(container, {
+    url: "/ws/terminal",
+    onClose(details) {
+      closed = details;
+      console.log("View closed", details.code, details.reason, details.wasClean);
+    }
+  });
+  terminal.focus();
+} catch (error) {
+  // closed is set for a transport close, but not for local initialization failure.
+  console.error("Mount failed", closed, error);
+}
+```
+
+Transport loss is **not workload completion**. Code 1006 means the browser did not
+receive a close frame; even code 1000 and `wasClean: true` only describe transport
+closure, not successful process exit. Hosts can define an application close-code
+contract and send it when their authoritative producer reports completion,
+including before any HWT frame exists. Hex1b does not assign workload meaning to
+close codes or reason strings. HTTP upgrade failures generally surface as 1006;
+browsers do not expose the rejected HTTP response body or status through this API.
+
+The client never retries automatically. The host decides whether to mount a new
+view after transport loss or leave an ended tab/dialog visible. Abort, explicit
+disposal, mount timeout, and local initialization/renderer failures do not
+synthesize `onClose`; no callback runs after disposal. Callback exceptions are
+reported to the host, not swallowed or retried, and do not leave mounting pending.
+The API does not retain the producer after exit or promise a final rendered frame.
+
 ### Browser and deployment requirements
 
 Use a browser with WebGPU or WebGL2, module workers, transferable OffscreenCanvas,
@@ -138,6 +183,7 @@ workers, fonts, and the intended WebSocket endpoint.
 | `readOnly` | Disable application input while retaining history inspection and selection. |
 | `label` | Accessible label for the terminal's hidden keyboard input. |
 | `onTitleChange` | Initial authoritative workload title, then distinct presented changes; see below. |
+| `onClose` | Native WebSocket close details, including pre-mount transport failure; not workload completion. |
 | `onProgressChange`, `onShellIntegrationChange` | Initial authoritative activity, then distinct presented changes for host-owned chrome. |
 | `inputBindings`, `onInput`, `actions` | Per-view input policy and custom actions. |
 | `onSelectionUI` | Synchronous, cancelable UI notification hook. |
