+++ src/web-terminal/src/wire-types.ts
@@ -1,6 +1,7 @@
 import type { InputModifiers, PointerButton, SelectionMode, SelectionRange,
   TerminalBuffer, TerminalFont, TerminalGeometry, TerminalPeer, TerminalSize, TerminalStats,
-  TerminalRendererPreference, TerminalStatusLevel, TerminalProgress, TerminalShellIntegration } from "./types.js";
+  TerminalRendererPreference, TerminalStatusLevel, TerminalProgress, TerminalShellIntegration,
+  TerminalCloseDetails } from "./types.js";
 
 export type SelectionText =
   | { status: "valid"; text: string }
@@ -81,7 +82,8 @@ export interface WorkerStats extends TerminalStats {
   history?: HistoryMetadata | null;
 }
 export type WorkerOutputMessage =
-  | { type: "connected" | "disconnected" }
+  | { type: "connected" }
+  | { type: "closed"; details: TerminalCloseDetails }
   | { type: "status"; message: string; level: TerminalStatusLevel }
   | ({ type: "geometry"; peer: TerminalPeer; history: HistoryMetadata | null;
        revision: number; title: string; progress: TerminalProgress; shellIntegration: TerminalShellIntegration;
